requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloc==0.2.5
clock==0.1
cmem-cmempy==20.12
cmempy==20.6.1
colored==1.4.2
configparser==5.0.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
greenlet==1.0.0
idna==2.10
isodate==0.6.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
neovim==0.3.1
pynvim==0.4.2
pyparsing==2.4.7
pytz==2021.1
rdflib==5.0.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
timeago==1.0.15
tzlocal==2.1
urllib3==1.26.3
